spi: spi-dw-mmio: Allow SPI core to manage bus number assignment
Commit 2418991efb7b1f9fc574771450ac17689682b3f6 allowed multiple
spi-dw-mmio instances by assigning the platform device ID to the
device instance bus number.  This inadvertently breaks the ability
of device tree users to set their desired bus number in the alias
DT node.

To allow the SPI core to assign bus numbers in the device tree case the
bus number must be negative when spi_register_master is called, so
clean up the only spi-dw user of the bus number, which is the
identifier used when requesting an IRQ, to use dev_name instead.  Also
remove dw_spi::name as there are no further users of this field.

With the cleanup done, set the DW-specific bus number to -1 by default
and assign the platform device ID to the bus number only in the non-DT
case.

Signed-off-by: Michael van der Westhuizen <michael@smart-africa.com>